Improves game creation and validation
Refactors the game creation process to utilize a DTO for input validation and data transfer.

This change ensures data integrity and provides a clear contract for the create game endpoint. It also allows for future expansion of game creation parameters.

Additionally, applies a global validation pipe to the application for request body validation.